+++ R/baseline.R
@@ -38,6 +38,16 @@
 #' @param cutoff Threshold for predicted classes. (Numeric)
 #'
 #'  N.B. \strong{Binomial only}
+#' @param random_generator_fn Function for generating random numbers when \code{type} is \code{"multinomial"}.
+#'  The softmax function is applied to the generated numbers to transform them to probabilities.
+#'
+#'  The first argument must be the number of random numbers to generate,
+#'  as no other arguments are supplied.
+#'
+#'  To test the effect of using different functions,
+#'  see \code{\link[cvms:multiclass_probability_tibble]{multiclass_probability_tibble}}.
+#'
+#'  N.B. \strong{Multinomial only}
 #' @param min_training_rows Minimum number of rows in the random subsets of \code{train_data}.
 #'
 #'  \strong{Gaussian only}. (Integer)
@@ -92,7 +102,8 @@
 #'
 #'  The \strong{Summarized Results} tibble contains:
 #'
-#'  Average \strong{RMSE}, \strong{MAE}, \strong{r2m}, \strong{r2c}, \strong{AIC}, \strong{AICc}, and \strong{BIC}.
+#'  Average \strong{RMSE}, \strong{MAE}, \strong{r2m}, \strong{r2c},
+#'  \strong{AIC}, \strong{AICc}, and \strong{BIC}.
 #'  }
 #'
 #'  The \strong{Measure} column indicates the statistical descriptor used on the evaluations.
@@ -232,6 +243,8 @@ baseline <- function(test_data,
                      # Binomial
                      positive = 2,
                      cutoff = 0.5,
+                     # Multinomial
+                     random_generator_fn = runif,
                      # Gaussian
                      min_training_rows = 5,
                      min_training_rows_left_out = 3,
@@ -265,7 +278,8 @@ baseline <- function(test_data,
       create_multinomial_baseline_evaluations(test_data = test_data,
                                               dependent_col = dependent_col,
                                               reps = n,
-                                              parallel_ = parallel
+                                              parallel_ = parallel,
+                                              random_generator_fn = random_generator_fn
       )
     )